test: run integration-test node from standalone node binaries
