+++ Cargo.toml
@@ -8,8 +8,8 @@ members = [
   "crates/rust-client",
   "crates/sqlite-store",
   "crates/testing/miden-client-tests",
-  "crates/testing/node-builder",
   "crates/testing/prover",
+  "crates/testing/test-node-genesis",
 ]
 
 default-members = ["bin/miden-cli", "crates/rust-client"]
@@ -42,18 +42,13 @@ miden-tx              = { default-features = false, version = "0.15" }
 miden-tx-batch-prover = { default-features = false, version = "0.15" }
 
 # Miden node dependencies
-miden-node-block-producer = { git = "https://github.com/0xMiden/node.git", rev = "6649a4ce774bc842c08e6bdc314f6ddafb816282" }
 miden-node-proto = { git = "https://github.com/0xMiden/node.git", rev = "6649a4ce774bc842c08e6bdc314f6ddafb816282" }
 miden-node-proto-build = { default-features = false, git = "https://github.com/0xMiden/node.git", rev = "6649a4ce774bc842c08e6bdc314f6ddafb816282" }
-miden-node-rpc = { git = "https://github.com/0xMiden/node.git", rev = "6649a4ce774bc842c08e6bdc314f6ddafb816282" }
-miden-node-store = { git = "https://github.com/0xMiden/node.git", rev = "6649a4ce774bc842c08e6bdc314f6ddafb816282" }
 miden-node-utils = { git = "https://github.com/0xMiden/node.git", rev = "6649a4ce774bc842c08e6bdc314f6ddafb816282" }
 miden-note-transport-proto-build = { default-features = false, version = "0.2" }
-miden-ntx-builder = { git = "https://github.com/0xMiden/node.git", rev = "6649a4ce774bc842c08e6bdc314f6ddafb816282" }
 miden-remote-prover-client = { default-features = false, features = [
   "tx-prover",
 ], git = "https://github.com/0xMiden/node.git", rev = "6649a4ce774bc842c08e6bdc314f6ddafb816282" }
-miden-validator = { git = "https://github.com/0xMiden/node.git", rev = "6649a4ce774bc842c08e6bdc314f6ddafb816282" }
 
 # Miden debug dependency
 miden-debug     = { default-features = false, features = ["dap", "std"], version = "0.8" }
