--- crates/testing/node-builder/Cargo.toml
@@ -1,46 +1,0 @@
-[package]
-authors.workspace      = true
-description            = "A minimal node implementation for running integration tests of the Miden client. NOT for production use."
-edition.workspace      = true
-keywords               = ["miden", "node", "protobuf", "rpc", "testing"]
-license.workspace      = true
-name                   = "node-builder"
-publish                = false
-readme                 = "README.md"
-repository.workspace   = true
-rust-version.workspace = true
-version.workspace      = true
-
-[[bin]]
-name = "testing-node-builder"
-path = "src/main.rs"
-
-[dependencies]
-# Miden dependencies
-miden-node-block-producer = { workspace = true }
-miden-node-proto          = { workspace = true }
-miden-node-rpc            = { workspace = true }
-miden-node-store          = { workspace = true }
-miden-node-utils          = { features = ["testing", "tracing-forest"], workspace = true }
-miden-ntx-builder         = { workspace = true }
-miden-protocol            = { workspace = true }
-miden-standards           = { workspace = true }
-miden-validator           = { workspace = true }
-
-# External dependencies
-anyhow      = { workspace = true }
-rand        = { workspace = true }
-rand_chacha = { workspace = true }
-tokio       = { features = ["signal"], workspace = true }
-tonic       = { workspace = true }
-url         = { features = ["serde"], version = "2.5" }
-
-[dev-dependencies]
-tempfile = { workspace = true }
-
-[lints]
-workspace = true
-
-[lib]
-doctest = false
-test    = false
