+++ crates/testing/test-node-genesis/Cargo.toml
@@ -1,0 +1,33 @@
+[package]
+authors.workspace      = true
+description            = "Generates genesis fixtures for bootstrapping a Miden client testing node. NOT for production use."
+edition.workspace      = true
+keywords               = ["genesis", "miden", "node", "testing"]
+license.workspace      = true
+name                   = "test-node-genesis"
+publish                = false
+readme                 = "README.md"
+repository.workspace   = true
+rust-version.workspace = true
+version.workspace      = true
+
+[[bin]]
+name = "gen-genesis"
+path = "src/bin/gen-genesis.rs"
+
+[dependencies]
+# Miden dependencies
+miden-protocol  = { features = ["std", "testing"], workspace = true }
+miden-standards = { features = ["std"], workspace = true }
+
+# External dependencies
+anyhow      = { workspace = true }
+rand        = { workspace = true }
+rand_chacha = { workspace = true }
+
+[lints]
+workspace = true
+
+[lib]
+doctest = false
+test    = false
